Move FrontendConstants to map.types.ts 	deleted:    src/interface/src/app/frontend-constants.ts 	modified:   src/interface/src/app/map/map-manager.ts 	modified:   src/interface/src/app/plan/plan-map/plan-map.component.ts 	modified:   src/interface/src/app/types/map.types.ts
