--- src/interface/src/app/frontend-constants.ts
@@ -1,6 +1,0 @@
-export const FrontendConstants = {
-  MAP_CENTER: [38.646, -120.548],
-  MAP_INITIAL_ZOOM: 9,
-  MAP_MIN_ZOOM: 7,
-  MAP_MAX_ZOOM: 13
-} as const;
