+++ src/interface/src/app/types/map.types.ts
@@ -54,3 +54,10 @@ export function defaultMapViewOptions(): MapViewOptions {
     center: [38.646, -120.548],
   };
 }
+
+export const FrontendConstants = {
+  MAP_CENTER: [38.646, -120.548],
+  MAP_INITIAL_ZOOM: 9,
+  MAP_MIN_ZOOM: 7,
+  MAP_MAX_ZOOM: 13
+} as const;
